// src/types/inventory/ItemSupplier.ts
export interface ItemSupplier {
  supplierName: any;
  itemSupplierId: number;
  name: string;
  phoneNumber: string;
  email: string;
  address: string;
  description: string;

}

export interface ItemSupplierFormData {
   supplierName: any;
  name: string;
  phoneNumber: string;
  email: string;
  address: string;
  description: string;
}